Move out posts module into top-lev modules
