+++ modules/post/client-react/components/PostEditView.jsx
@@ -2,9 +2,9 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { StyleSheet, View, ScrollView, Keyboard } from 'react-native';
 import KeyboardSpacer from 'react-native-keyboard-spacer';
+import { translate } from '@module/i18n-client-react';
+import { Loading } from '@module/look-client-react-native';
 
-import { Loading } from '../../common/components/native';
-import translate from '../../../i18n';
 import PostForm from './PostForm';
 import PostComments from '../containers/PostComments';
