+++ modules/post/client-react/components/PostForm.jsx
@@ -2,12 +2,11 @@ import React from 'react';
 import { View } from 'react-native';
 import PropTypes from 'prop-types';
 import { withFormik } from 'formik';
-
-import translate from '../../../i18n';
-import Field from '../../../utils/FieldAdapter';
-import { FormView, RenderField, Button, primary } from '../../common/components/native';
-import { placeholderColor, submit } from '../../common/components/native/styles';
-import { required, validate } from '../../../../../common/modules/validation';
+import { translate } from '@module/i18n-client-react';
+import { FieldAdapter as Field } from '@module/core-client-react';
+import { required, validate } from '@module/validation-common-react';
+import { FormView, RenderField, Button, primary } from '@module/look-client-react-native';
+import { placeholderColor, submit } from '@module/look-client-react-native/styles';
 
 const postFormSchema = {
   title: [required],
