+++ modules/post/client-react/components/PostForm.web.jsx
@@ -1,11 +1,10 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { withFormik } from 'formik';
-
-import translate from '../../../i18n';
-import Field from '../../../utils/FieldAdapter';
-import { Form, RenderField, Button } from '../../common/components/web';
-import { required, validate } from '../../../../../common/modules/validation';
+import { translate } from '@module/i18n-client-react';
+import { FieldAdapter as Field } from '@module/core-client-react';
+import { required, validate } from '@module/validation-common-react';
+import { Form, RenderField, Button } from '@module/look-client-react';
 
 const postFormSchema = {
   title: [required],
