+++ modules/post/client-react/graphql/AddPost.graphql
