+++ modules/post/client-react/graphql/DeleteComment.graphql
