+++ modules/post/client-react/graphql/DeletePost.graphql
