+++ modules/post/client-react/graphql/PostSubscription.graphql
