+++ modules/post/client-react/graphql/PostsQuery.graphql
