+++ modules/post/client-react/graphql/PostsSubscription.graphql
