+++ modules/post/client-react/index.jsx
@@ -1,14 +1,14 @@
 import React from 'react';
 import { Route, NavLink } from 'react-router-dom';
-import translate from '../../i18n';
+import { translate } from '@module/i18n-client-react';
+import { MenuItem } from '@module/look-client-react';
+import ClientModule from '@module/module-client-react';
 
-import { MenuItem } from '../../modules/common/components/web';
 import Post from './containers/Post';
 import PostEdit from './containers/PostEdit';
 import PostAdd from './containers/PostAdd';
 import resources from './locales';
 import resolvers from './resolvers';
-import ClientModule from '../ClientModule';
 
 const NavLinkWithI18n = translate()(({ t }) => (
   <NavLink to="/posts" className="nav-link" activeClassName="active">
