+++ modules/post/client-react/locales/index.js
