+++ modules/post/client-react/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "@module/post-client-react",
+  "version": "1.0.0",
+  "private": true
+}
