+++ modules/post/client-react/resolvers/index.js
