+++ packages/client/src/index.tsx
@@ -9,10 +9,10 @@ import ClientModule from '@module/module-client-react';
 import defaultRouter from '@module/router-client-react';
 import '@module/favicon-common';
 
-import post from './modules/post';
 import user from './modules/user';
 import payments from './modules/payments';
 
+const post = require('@module/post-client-react').default;
 const pageNotFound = require('@module/page-not-found-client-react');
 const upload = require('@module/upload-client-react').default;
 const pagination = require('@module/pagination-client-react').default;
