+++ packages/mobile/src/index.ts
@@ -6,12 +6,12 @@ import contact from '@module/contact-client-react';
 import validation from '@module/validation-common-react';
 import defaultRouter from '@module/router-client-react-native';
 
-import post from '../../client/src/modules/post';
 import user from '../../client/src/modules/user';
 import payments from '../../client/src/modules/payments';
 
 import ClientModule from '@module/module-client-react-native';
 
+const post = require('@module/post-client-react').default;
 const upload = require('@module/upload-client-react').default;
 const pagination = require('@module/pagination-client-react').default;
